refactor(api): move JWT tokens and security utilities to core directory
- Moved jwt_tokens.py from utils/ to core/ directory
- Moved security.py from utils/ to core/ directory
- Updated all import statements across the codebase
- Maintained all existing functionality while improving code organization

This refactoring improves the project structure by separating core authentication
and security utilities into a dedicated core module, following better architectural patterns.
