+++ api/services/auth_service.py
@@ -6,17 +6,17 @@ from typing import Any
 import jwt
 from sqlalchemy.orm import Session
 
-from models import RefreshToken, User
-from utils.config import settings
-from utils.jwt_tokens import (
+from core.jwt_tokens import (
     TokenExpiredError,
     TokenInvalidError,
     create_access_token,
     create_refresh_token,
     verify_token,
 )
+from core.security import verify_password
+from models import RefreshToken, User
+from utils.config import settings
 from utils.logging import get_logger
-from utils.security import verify_password
 
 logger = get_logger()
